requirements.txt
google-adk
litellm
pandas
matplotlib
seaborn
pillow
numpy
scipy
scikit-learn
